fix(calendar): make preview date and selected date(s) tabbable (#1230)
Applies tabindex="0" to the calendar day cell that matches the
previewDate value as well as the one(s) marked selected. Since the
previewDate always matches a cell within the currently visible months
this ensures that the calendar stays in the tab order even if the
current selection is not within the visible month(s).